java-berlin-clock/src/main/java/com/ubs/opsit/interviews/model/Clock.java
package com.ubs.opsit.interviews.model;

public class Clock {

    private String hours;

    private String minutes;

    private String seconds;

    public Clock() {
    }

    public Clock(String hours, String minutes, String seconds) {
        this.hours = hours;
        this.minutes = minutes;
        this.seconds = seconds;
    }

    public String getHours() {
        return hours;
    }

    public void setHours(String hours) {
        this.hours = hours;
    }

    public String getMinutes() {
        return minutes;
    }

    public void setMinutes(String minutes) {
        this.minutes = minutes;
    }

    public String getSeconds() {
        return seconds;
    }

    public void setSeconds(String seconds) {
        this.seconds = seconds;
    }
}
